Found a fatal flaw in the combination of precompilation, module reinstantiation, and expression refs.
The issue is that a precompiled function that uses something like qs[] will have expression-refs to its values. These expression refs are synthesized as closure variables, which are then used
to compile the function into a live Caterwaul module. However, when that module is then re-serialized, those expression refs appear to be opaque variables and their values are lost. There
needs to be some way to save expression refs' values along with the live value they are used for. (This live value is most likely a function.)